Update version to 0.2.2; replace mockup image with demo GIF in README files; fix bugs caused by CORS and unnecessary auth check in web-clipper API.
