+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "knowlink",
-  "version": "0.2.1",
+  "version": "0.2.2",
   "private": true,
   "scripts": {
     "bootstrap": "bun ./scripts/bootstrap.ts",
