+++ src/app/(main)/components/side-bar/main-menu.tsx
@@ -86,7 +86,7 @@ export function MainMenu() {
                   title: "About Knowlink",
                   body: (
                     <div>
-                      <p>Version: 0.2.1</p>
+                      <p>Version: 0.2.2</p>
                       <p>
                         Website:{" "}
                         <a
